+++ lib/apathy_drive/data/stock/commands/attack.ex
@@ -10,7 +10,7 @@ defmodule Commands.Attack do
   def execute(_spirit, monster, arguments) do
     current_room = Parent.of(monster)
 
-    target = current_room |> find_entity_in_room(Enum.join(arguments, " "))
+    target = current_room |> find_entity_in_room(monster, Enum.join(arguments, " "))
     attack(monster, target)
   end
 
@@ -18,18 +18,15 @@ defmodule Commands.Attack do
     send_message(entity, "scroll", "<p>Attack what?</p>")
   end
 
-  defp attack(entity, target) when entity == target do
-    send_message(entity, "scroll", "<p>Attack yourself?</p>")
-  end
-
   defp attack(entity, target) do
     send_message(entity, "scroll", "<p><span class='dark-yellow'>You move to attack #{Components.Name.value(target)}!</span></p>")
     Systems.Combat.attack(entity, target)
   end
 
-  defp find_entity_in_room(room, string) do
+  defp find_entity_in_room(room, monster, string) do
     room
     |> Systems.Room.living_in_room
+    |> Enum.reject(&(&1 == monster))
     |> Systems.Match.one(:name_contains, string)
   end