+++ lib/apathy_drive/systems/skill.ex
@@ -57,7 +57,7 @@ defmodule Systems.Skill do
 
           average = total / (Map.values(modifiers) |> Enum.sum)
 
-          round(base * (1 + average * 0.02))
+          round(base * (1 + average * 0.005))
         else
           0
         end